refactor: regroup all types in a single types.d.ts file
